added modal for setting and uploadGiffy
