+++ webapp/API/serverHooks.ts
@@ -1,59 +1,71 @@
 import { giffyDTO } from './DTO';
 
 export async function getCollectionsByUserId(userId: number) {
-	const response = await fetch(
-		`${process.env.NEXT_PUBLIC_SERVER_URL}/collections/getCollectionsByUserId/${userId}`,
-		{
-			mode: 'cors',
-			cache: 'no-cache',
-			credentials: 'same-origin',
-			headers: {
-				'Content-Type': 'application/json',
-			},
-			method: 'GET',
-		}
-	);
+	try {
+		const response = await fetch(
+			`${process.env.NEXT_PUBLIC_SERVER_URL}/collections/getCollectionsByUserId/${userId}`,
+			{
+				mode: 'cors',
+				cache: 'no-cache',
+				credentials: 'same-origin',
+				headers: {
+					'Content-Type': 'application/json',
+				},
+				method: 'GET',
+			}
+		);
 
-	return response.json();
+		return response.json();
+	} catch (e) {
+		console.log(e);
+	}
 }
 
 export async function createGiffy(data: {
 	collectionId: number;
 	firebaseUrl: string;
 	giffyName: string;
-}): Promise<giffyDTO> {
-	const response = await fetch(
-		`${process.env.NEXT_PUBLIC_SERVER_URL}/giffies/createGiffy`,
-		{
-			method: 'POST',
-			mode: 'cors',
-			cache: 'no-cache',
-			credentials: 'same-origin',
-			headers: {
-				'Content-Type': 'application/json',
-			},
-			body: JSON.stringify(data),
-		}
-	);
+}) {
+	try {
+		const response = await fetch(
+			`${process.env.NEXT_PUBLIC_SERVER_URL}/giffies/createGiffy`,
+			{
+				method: 'POST',
+				mode: 'cors',
+				cache: 'no-cache',
+				credentials: 'same-origin',
+				headers: {
+					'Content-Type': 'application/json',
+				},
+				body: JSON.stringify(data),
+			}
+		);
 
-	return response.json();
+		return response.json();
+	} catch (err) {
+		console.log(e);
+	}
 }
 
 export async function getUserByFirebaseAuthId(firebaseAuthId: string) {
-	const response = await fetch(
-		`${process.env.NEXT_PUBLIC_SERVER_URL}/users/getUserByFirebaseAuthId/${firebaseAuthId}`,
-		{
-			mode: 'cors',
-			cache: 'no-cache',
-			credentials: 'same-origin',
-			headers: {
-				'Content-Type': 'application/json',
-			},
-			method: 'GET',
-		}
-	);
+	try {
+		const response = await fetch(
+			`${process.env.NEXT_PUBLIC_SERVER_URL}/users/getUserByFirebaseAuthId/${firebaseAuthId}`,
+			{
+				mode: 'cors',
+				cache: 'no-cache',
+				credentials: 'same-origin',
+				headers: {
+					'Content-Type': 'application/json',
+				},
+				method: 'GET',
+			}
+		);
 
-	return response.json();
+		return response.json();
+	} catch (e) {
+		console.log(e);
+	}
 }
 
 export async function createCollection(data: {
@@ -61,36 +73,44 @@ export async function createCollection(data: {
 	private: boolean;
 	userId: number;
 }) {
-	const response = await fetch(
-		`${process.env.NEXT_PUBLIC_SERVER_URL}/collections/createCollection`,
-		{
-			method: 'POST',
-			mode: 'cors',
-			cache: 'no-cache',
-			credentials: 'same-origin',
-			headers: {
-				'Content-Type': 'application/json',
-			},
-			body: JSON.stringify(data),
-		}
-	);
+	try {
+		const response = await fetch(
+			`${process.env.NEXT_PUBLIC_SERVER_URL}/collections/createCollection`,
+			{
+				method: 'POST',
+				mode: 'cors',
+				cache: 'no-cache',
+				credentials: 'same-origin',
+				headers: {
+					'Content-Type': 'application/json',
+				},
+				body: JSON.stringify(data),
+			}
+		);
 
-	return response.json();
+		return response.json();
+	} catch (e) {
+		console.log(e);
+	}
 }
 
 export async function getGiffiesByCollectionId(collectionId: number) {
-	const response = await fetch(
-		`${process.env.NEXT_PUBLIC_SERVER_URL}/giffies/getGiffiesByCollectionId/${collectionId}`,
-		{
-			mode: 'cors',
-			cache: 'no-cache',
-			credentials: 'same-origin',
-			headers: {
-				'Content-Type': 'application/json',
-			},
-			method: 'GET',
-		}
-	);
+	try {
+		const response = await fetch(
+			`${process.env.NEXT_PUBLIC_SERVER_URL}/giffies/getGiffiesByCollectionId/${collectionId}`,
+			{
+				mode: 'cors',
+				cache: 'no-cache',
+				credentials: 'same-origin',
+				headers: {
+					'Content-Type': 'application/json',
+				},
+				method: 'GET',
+			}
+		);
 
-	return response.json();
+		return response.json();
+	} catch (e) {
+		console.log(e);
+	}
 }
