+++ webapp/pages/collections/[collection].tsx
@@ -1,17 +1,14 @@
 import type { NextPage } from 'next';
 import Layout from '../../components/Layout';
-import Link from 'next/link';
 import Card from '../../components/Card';
 import { useEffect, useState } from 'react';
-import { getGiffiesByCollectionId } from '../../API/serverHooks';
 import { giffyDTO } from '../../API/DTO';
 import CardDistributor from '../../components/CardDistributor';
-import styles from '../../styles/Collections.module.scss';
 import { useRouter } from 'next/router';
 import { useAppDispatch, useAppSelector } from '../../hooks';
 import { RootState } from '../../store';
-import UploadGiffy from '../../components/UploadGiffy';
-import layoutStyles from '../../styles/Layout.module.scss';
+import styles from '../../styles/Collections.module.scss';
+
 import { openUploadGiffyWindow } from '../../slices/CollectionsSlice';
 
 const Collections: NextPage = () => {
@@ -19,9 +16,6 @@ const Collections: NextPage = () => {
 	const router = useRouter();
 	const { collection } = router.query;
 	const dispatch = useAppDispatch();
-	const isUploadGiffyWindowOpen = useAppSelector(
-		(state: RootState) => state.collections.isUploadGiffyWindowOpen
-	);
 	const giffies = useAppSelector((state: RootState) => {
 		return state.collections.value?.filter(
 			curCollection => curCollection.collectionId === Number(collection)
@@ -59,11 +53,6 @@ const Collections: NextPage = () => {
 			) : (
 				<p>No items yet</p>
 			)}
-			{isUploadGiffyWindowOpen ? (
-				<div className={layoutStyles.settingWindow}>
-					<UploadGiffy currentCollectionId={Number(collection)} />
-				</div>
-			) : null}
 		</Layout>
 	);
 };
