--- webapp/styles/AccountSettings.module.scss
@@ -1,39 +1,0 @@
-.accountSettingsWindow {
-	width: 100%;
-	height: 100%;
-	background-color: #f0f2f3;
-	padding: 17px;
-	overflow: auto;
-
-	// font-family: Arial, Helvetica, sans-serif,helvetica,arial,sans-serif;
-	// line-height: 1.4;
-	// font-size: 14px;
-	// margin: 0;
-	// padding: 0;
-	// border: 0;
-	// text-align: left;
-	// background: #FFF;
-	// z-index: 777;
-	// width: 1306px;
-	// top: 50px;
-	// position: fixed;
-	// left: 100px;
-	// display: block;
-	// text-align: left;
-
-	.closeIcon {
-		content: 'x';
-		cursor: pointer;
-		position: absolute;
-		top: 11px;
-		right: 11px;
-		background: #ededed;
-		width: 25px;
-		height: 25px;
-		border-radius: 70%;
-		line-height: 20px;
-		text-align: center;
-		border: 1px solid #999;
-		font-size: 20px;
-	}
-}
