+++ webapp/styles/Card.module.scss
@@ -19,12 +19,12 @@
 		width: 100%;
 		border-radius: 5px 5px 0 0;
 	}
+
+	.container {
+		padding: 2px 16px;
+	}
 }
 
 .card:hover {
 	box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.4);
-}
-
-.container {
-	padding: 2px 16px;
 }
