+++ webapp/styles/SidePanel.module.scss
@@ -1,23 +1,30 @@
 .sidePanel {
-	width: 25%;
 	position: relative;
-	z-index: 2;
 	background-color: #6e9b7b;
 	transition: 0.5s ease;
 	overflow-x: visible;
 	padding-top: 20px;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
 
-	a {
-		color: white;
-		width: 100%;
-		justify-content: center;
-		padding: 0.5em;
-		text-align: center;
-		&:hover {
-			background-color: #2f2e2e;
+	.sidePanelContent {
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+		overflow-x: hidden;
+
+		a {
+			color: white;
+			width: 100%;
+			justify-content: center;
+			padding: 0.5em;
+			text-align: center;
+			&:hover {
+				background-color: #2f2e2e;
+			}
+		}
+
+		// to clear the bullet point before link
+		ul {
+			list-style-type: none;
 		}
 	}
 
@@ -31,7 +38,6 @@
 		cursor: pointer;
 
 		// position on the edge of the side nav
-		z-index: 9999;
 		position: absolute;
 		right: -20px;
 
@@ -43,10 +49,5 @@
 		&:hover {
 			background-color: RoyalBlue;
 		}
-	}
-
-	// to clear the bullet point before link
-	ul {
-		list-style-type: none;
 	}
 }
